server/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.4
pandas==2.3.1
matplotlib==3.9.4
pyTelegramBotAPI==4.28.0
tiktoken==0.11.0
pydantic==2.11.7
google-genai==1.52.0
httpx
claude-agent-sdk
